account_invoice_triple_discount: Some improvements
