+++ account_invoice_triple_discount/models/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 from . import account_invoice
-from . import account_invoice_line
